+++ tests/context.py
@@ -7,7 +7,7 @@ sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "../u
 # Import the required classes and functions
 # pylint: disable=unused-import,wrong-import-position
 from up42.tools import Tools
-from up42.visualization import folium_base_map, Visualization
+from up42.viztools import folium_base_map, VizTools
 from up42.utils import (
     any_vector_to_fc,
     fc_to_query_geometry,
