+++ up42/catalog.py
@@ -13,7 +13,7 @@ from tqdm import tqdm
 
 from up42.auth import Auth
 from up42.tools import Tools
-from up42.visualization import Visualization
+from up42.viztools import VizTools
 from up42.utils import get_logger, any_vector_to_fc, fc_to_query_geometry
 
 logger = get_logger(__name__)
@@ -59,7 +59,7 @@ supported_sensors = {
 }
 
 # pylint: disable=duplicate-code
-class Catalog(Visualization, Tools):
+class Catalog(VizTools, Tools):
     def __init__(self, auth: Auth):
         """
         The Catalog class enables access to the UP42 catalog search. You can search
